+++ UnityUtils/Assets/Testing Scripts/TestingScript3.cs
@@ -1,5 +1,5 @@
 using UnityEngine;
-using UnityUtils.ScriptUtils.Objects;
+using UnityUtils.ScriptUtils.Objects.ColorFlash;
 
 public class TestingScript3 : MonoBehaviour {
   public GameObject mat3;